+++ src/main/java/cart/product/application/ProductService.java
@@ -1,13 +1,13 @@
-package cart.application;
+package cart.product.application;
 
-import cart.controller.dto.ProductRequest;
-import cart.controller.dto.ProductResponse;
-import cart.controller.dto.ProductsResponse;
-import cart.domain.Product;
-import cart.domain.ProductRepository;
-import cart.domain.Products;
 import cart.exception.ErrorType;
 import cart.exception.ServiceException;
+import cart.product.controller.dto.ProductRequest;
+import cart.product.controller.dto.ProductResponse;
+import cart.product.controller.dto.ProductsResponse;
+import cart.product.domain.Product;
+import cart.product.domain.ProductRepository;
+import cart.product.domain.Products;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -32,21 +32,21 @@ public class ProductService {
         return ProductResponse.of(product);
     }
 
-    public ProductResponse updateProduct(long id, ProductRequest request) {
-        Product product = findProduct(id);
+    public ProductResponse updateProduct(long productId, ProductRequest request) {
+        Product product = findProduct(productId);
         product.update(request.getName(), request.getImage(), request.getPrice());
         productRepository.update(product);
 
         return ProductResponse.of(product);
     }
 
-    private Product findProduct(long id) {
-        return productRepository.findById(id)
+    private Product findProduct(long productId) {
+        return productRepository.findById(productId)
                 .orElseThrow(() -> new ServiceException(ErrorType.PRODUCT_NOT_FOUND));
     }
 
-    public void deleteProduct(long id) {
-        Product product = findProduct(id);
+    public void deleteProduct(long productId) {
+        Product product = findProduct(productId);
         productRepository.delete(product);
     }
