+++ src/main/java/cart/product/controller/dto/ProductRequest.java
@@ -1,6 +1,6 @@
-package cart.controller.dto;
+package cart.product.controller.dto;
 
-import cart.domain.Product;
+import cart.product.domain.Product;
 
 public class ProductRequest {
