+++ src/main/java/cart/product/domain/ProductRepository.java
@@ -1,5 +1,6 @@
-package cart.domain;
+package cart.product.domain;
 
+import java.util.List;
 import java.util.Optional;
 
 public interface ProductRepository {
@@ -14,4 +15,5 @@ public interface ProductRepository {
 
     void delete(Product product);
 
+    Products findByIds(List<Long> productIds);
 }
