+++ src/main/java/cart/product/domain/Products.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.product.domain;
 
 import java.util.List;
