+++ src/main/java/cart/product/infra/ProductDao.java
@@ -1,9 +1,8 @@
-package cart.infra;
+package cart.product.infra;
 
-import cart.domain.Product;
-import cart.domain.ProductRepository;
-import cart.domain.Products;
-import cart.infra.rowmapper.ProductRowMapper;
+import cart.product.domain.Product;
+import cart.product.domain.ProductRepository;
+import cart.product.domain.Products;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcDaoSupport;
@@ -74,6 +73,15 @@ public class ProductDao extends NamedParameterJdbcDaoSupport implements ProductR
         SqlParameterSource namedParameters = new MapSqlParameterSource().addValue("id", product.getId());
 
         getNamedParameterJdbcTemplate().update(query, namedParameters);
+    }
+
+    @Override
+    public Products findByIds(List<Long> productIds) {
+        final String query = String.format("SELECT * FROM %s WHERE id IN (:ids)", TABLE_NAME);
+        SqlParameterSource namedParameters = new MapSqlParameterSource().addValue("ids", productIds);
+        List<Product> products = getNamedParameterJdbcTemplate().query(query, namedParameters, ROW_MAPPER);
+
+        return new Products(products);
     }
 
 }
