+++ src/main/java/cart/product/infra/ProductRowMapper.java
@@ -1,7 +1,7 @@
-package cart.infra.rowmapper;
+package cart.product.infra;
 
-import cart.domain.Money;
-import cart.domain.Product;
+import cart.product.domain.Money;
+import cart.product.domain.Product;
 import org.springframework.jdbc.core.RowMapper;
 
 import java.sql.ResultSet;
